requirements.txt
Django==4.2
django-cors-headers==3.13.0
django-environ==0.9.0
cachetools==5.2.0
certifi==2022.9.24
django-crispy-forms
crispy-bootstrap5
pillow==10.3.0
Jinja2==3.1.3
phonenumbers==8.13.39
django-countries==7.6.1
django-admin-tools
django-admin-charts